Moved location of `setLoading(false)`
Moved location of `setLoading(false)` to inside of the videoConferenceJoined event listener.  If loading is set to false before the conference is joined, the loading component will only flash for a split second and then there will be a blank screen until the conference is joined.  This way, the loading component will be shown until the conference is ready to go.