Add password required event listener
If a password (prop) is set when creating a room, the password is set in the videoConferenceJoined event listener.  If someone is joining a room that requires a password, it's either entered automatically (if the password prop was set) or the setLoading(false) in the passwordRequired event listener will allow them to see the "enter password" UI prompt.